Communication
Agents need to know about each other, so they all need the list of agents.
Using the neighbourhood method,
the agents scan the list of other agents, and if they find any within a neighbourhood of themselves, they adjust their variables and their neighbours' variables in order to transfer information to the other agents.