Environment control
Control on the amount of environment that can be eaten is added to the model.
A grazing limit of 60000 units is added to control the amount of environment that can be eaten. Once the agents eat the amount from the environment, they stop moving to eat around but they can still breed among themselves.